move wazuh access to ec2/security-groups.ts
