+++ src/constructs/ec2/security-groups.ts
@@ -1,5 +1,5 @@
 import type { CfnSecurityGroup, IPeer, SecurityGroupProps } from "@aws-cdk/aws-ec2";
-import { Port, SecurityGroup } from "@aws-cdk/aws-ec2";
+import { Peer, Port, SecurityGroup } from "@aws-cdk/aws-ec2";
 import type { GuStack } from "../core";
 
 export interface CidrIngress {
@@ -36,5 +36,26 @@ export class GuSecurityGroup extends SecurityGroup {
     props.egresses?.forEach(({ range, description, port }) =>
       this.addEgressRule(range, port, description ?? undefined)
     );
+  }
+}
+
+export class GuWazuhAccess extends GuSecurityGroup {
+  private static getDefaultProps(): Partial<GuSecurityGroupProps> {
+    return {
+      description: "Wazuh agent registration and event logging",
+      overrideId: true,
+      allowAllOutbound: false,
+      egresses: [
+        { range: Peer.anyIpv4(), port: Port.tcp(1514), description: "wazuh event logging" },
+        { range: Peer.anyIpv4(), port: Port.tcp(1515), description: "wazuh agent registration" },
+      ],
+    };
+  }
+
+  constructor(scope: GuStack, id: string = "WazuhSecurityGroup", props: GuSecurityGroupProps) {
+    super(scope, id, {
+      ...GuWazuhAccess.getDefaultProps(),
+      ...props,
+    });
   }
 }
